// src/integration/api-types.ts
export interface components {
    schemas: {
// ...
        GemeenteDetail: {
            id: string;
            /** @example Amsterdam */
            naam: string;
            /** @example 0363 */
            cbsCode: string;
            /** @example 85 */
            progress: number;
            contactpersoon?: string | null;
            email?: string | null;
            website?: string | null;
            telefoon?: string | null;
            samenwerkingen: {
                /** @example Dimpact */
                naam?: string;
                /** @example Coöperatie */
                type?: string;
            }[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}